backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart

langchain
langgraph
ollama

easyocr
opencv-python
faster-whisper

sentence-transformers
faiss-cpu
numpy

sympy

python-dotenv
tqdm